+++ doc/CHANGELOG.rst
@@ -5,6 +5,9 @@ This changelog *only* contains changes from the *first* pypi release (0.5.4.3) o
 
 Latest Changes:
 - **1.4.1_dev0 - 2022-05-14**
+
+  - Launch JVM in an independent thread.
+
 - **1.4.0 - 2022-05-14**
 
   - Support for all different buffer type conversions.